Move inflators out of jsx-entity.js
